Update guessing game to version 1.0.1 and guess type to u32
Increase version of the guessing_game crate from 1.0.0 to 1.0.1 in both Cargo.lock and Cargo.toml files. Additionally, change guessing variable type from u8 to u32 in the main.rs file.

+++ src/main.rs
@@ -20,7 +20,7 @@ fn main() {
             .read_line(&mut guess)
             .expect("Failed to read line");
 
-        let guess: u8 = match guess.trim().parse() {
+        let guess: u32 = match guess.trim().parse() {
             Ok(num) => num,
             Err(_) => continue,
         };
